Crate sgx_tse

Trusted SE Library

The library provides functions for getting specific keys and for creating and verifying an enclave report.

Functions

rsgx_create_report
The rsgx_create_report function tries to use the information of the target enclave and other information to create a cryptographic report of the enclave.

rsgx_get_key
The rsgx_get_key function generates a 128-bit secret key using the input information.

rsgx_verify_report
The rsgx_verify_report function provides software verification for the report which is expected to be generated by the rsgx_create_report function.