Function sgx_tse::rsgx_verify_report

pub fn rsgx_verify_report(report: &sgx_report_t) -> SgxError

The rsgx_verify_report function provides software verification for the report which is expected to be generated by the rsgx_create_report function.

Description

The rsgx_verify_report performs a cryptographic CMAC function of the input sgx_report_data_t object in the report using the report key. Then the function compares the input report MAC value with the calculated MAC value to determine whether the report is valid or not.

Parameters

report

A pointer to an sgx_report_t object that contains the cryptographic report to be verified. The report buffer must be within the enclave.

Requirements

Library: libsgx_tservice.a

Errors

SGX_ERROR_INVALID_PARAMETER

The report object is invalid.

SGX_ERROR_MAC_MISMATCH

Indicates report verification error.

SGX_ERROR_UNEXPECTED

Indicates an unexpected error occurs during the report verification process.